상태 변경하는 부분을 store에서 reducer.js로 분리
