+++ src/reducer.js
@@ -1,0 +1,39 @@
+const initalState = {
+  newId: 100,
+  taskTitle: '',
+  tasks: [
+    { id: 1, title: '아무 것도 하지 않기 #1' },
+    { id: 2, title: '아무 것도 하지 않기 #2' },
+  ],
+};
+
+export default function reducer(state = initalState, action) {
+  if (action.type === 'updateTaskTitle') {
+    return {
+      ...state,
+      taskTitle: action.payload.taskTitle,
+    };
+  }
+
+  if (action.type === 'addTask') {
+    const { newId, taskTitle, tasks } = state;
+
+    return {
+      ...state,
+      newId: newId + 1,
+      taskTitle: '',
+      tasks: [...tasks, { id: newId, title: taskTitle }],
+    };
+  }
+
+  if (action.type === 'deleteTask') {
+    const { tasks } = state;
+
+    return {
+      ...state,
+      tasks: tasks.filter((task) => task.id !== action.payload.id),
+    };
+  }
+
+  return state;
+}
